# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Go TDD Server Boilerplate (AI-Ready with MCP)",
    "version": "0.0.2",
    "description": "Professional Go API with TDD, Clean Architecture, MCP (AI Tools) and OIDC/JWKS Security."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide a valid JWT token issued by your OIDC provider (e.g., Authentik)."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/mcp/sse": {
      "get": {
        "summary": "MCP SSE Endpoint",
        "description": "Entry point for AI agents to connect via Model Context Protocol.",
        "responses": {
          "200": { "description": "SSE Event Stream" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}